requirements.txt
pandas==2.2.0
python-dotenv==1.0.0
tiktoken==0.7.0
google-generativeai==0.3.1
anthropic==0.30.0
matplotlib==3.8.3
seaborn==0.13.1
openai==1.10.0
openpyxl==3.1.2
xlrd==2.0.1
numpy==1.26.4
cohere==4.47
scikit-learn==1.4.0 